requirements.txt
pyarrow
pandas
python-dotenv
boto3
pyspark
apache-airflow